SaveManager Logic.

+++ src/edu/neumont/csc150/controller/SaveManager.java
@@ -6,8 +6,7 @@
  */
 package edu.neumont.csc150.controller;
 
-import edu.neumont.csc150.model.Game;
-import edu.neumont.csc150.model.Player;
+import edu.neumont.csc150.model.*;
 
 import java.io.*;
 import java.util.ArrayList;
@@ -32,31 +31,71 @@ public class SaveManager {
         writeSave(saveName, saveInfo.toString());
     }
 
-    private static void writeSave(String saveName, String saveInfo){
-        BufferedWriter bWriter;
-        try{
-            bWriter = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(getFullFilePath(saveName))));
-            try{
-                bWriter.write(saveInfo);
-            } catch(IOException ioe){
-                throw new RuntimeException(ioe);
-            } finally{
-                bWriter.close();
+    public static Game loadGame(String saveName){
+        List<String> saveData = readSave(saveName);
+        Game game = null;
+        Figure player = null;
+
+        if(saveData != null){
+            String playerName = null;
+            int trial = 0;
+            int health = Player.MAX_HEALTH;
+            int maxHealth = Player.MAX_HEALTH;
+            int damage = Player.BASE_DAMAGE;
+            Arsenal weapon = null;
+            for(String saveLine : saveData){
+                if(saveLine.startsWith("name")){
+                    String[] nameParts = saveLine.split(":");
+                    playerName = nameParts[1];
+                } else if(saveLine.startsWith("trial")){
+                    String[] trialParts = saveLine.split(":");
+                    trial = Integer.parseInt(trialParts[1]);
+                } else if(saveLine.startsWith("health")){
+                    String[] healthParts = saveLine.split(":");
+                    health = Integer.parseInt(healthParts[1]);
+                } else if(saveLine.startsWith("max health")){
+                    String[] maxHealthParts = saveLine.split(":");
+                    maxHealth = Integer.parseInt(maxHealthParts[1]);
+                }  else if(saveLine.startsWith("weapon")){
+                    String[] weaponParts = saveLine.split(":");
+                    weapon = Arsenal.valueOf(weaponParts[1]);
+                }
             }
-        }catch(IOException ioe){
-            ioe.printStackTrace();
+            player = new Player(playerName, weapon, damage, health, maxHealth);
+            game = new Game(player, trial);
+        }
+        return game;
+    }
+
+    private static void writeSave(String saveName, String saveInfo){
+        if(checkIfPathExists()) {
+            BufferedWriter bWriter;
+            try {
+                bWriter = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(getFullFilePath(saveName))));
+                try {
+                    bWriter.write(saveInfo);
+                } catch (IOException ioe) {
+                    throw new RuntimeException(ioe);
+                } finally {
+                    bWriter.close();
+                }
+            } catch (IOException ioe) {
+                ioe.printStackTrace();
+            }
         }
     }
     private static List<String> readSave(String saveName){
-        try{
-            BufferedReader bReader = new BufferedReader(new InputStreamReader(new FileInputStream(getFullFilePath(saveName))));
-            List<String> saveLines = new ArrayList<>();
-            while(bReader.ready()){
-                saveLines.add(bReader.readLine());
+        if(checkIfPathExists()) {
+            try {
+                BufferedReader bReader = new BufferedReader(new InputStreamReader(new FileInputStream(getFullFilePath(saveName))));
+                List<String> saveLines = new ArrayList<>();
+                while (bReader.ready()) {
+                    saveLines.add(bReader.readLine());
+                }
+                return saveLines;
+            } catch (IOException ioe) {
+                ioe.printStackTrace();
             }
-            return saveLines;
-        } catch(IOException ioe){
-            ioe.printStackTrace();
         }
         return null;
     }
